requirements.txt
discord.py[voice]
python-dotenv
jishaku
Motor
gTTS
ffmpeg